Improved CMake check for CAPSimage library
As well as checking for the library and main header, we now
check for the newest function we need from it (CAPSSetRevolution).
This avoids a linker error with CAPSimage 5.0, and a confusing
"CAPSimage not found" message if an older version is installed.
4.x and older didn't use the CapsLibAll.h header we look for.

Also added -Wsuggest-override if it's supported, to suggest
places we should be using the override keyword on overridden
virtual functions. This helps spot signature mismatches.