Changed to no longer override CRC status
If there's enough data in the data field to cover the CRC bytes
we check the CRC status is consistent with its value. Previously
we'd override the status, but now we simply check it (and assert).

This shouldn't happen and is usually an indication the source is
invalid in some way. This may happen if the sector uses a
non-standard DAM, which forms part of the CRC.