Added indicator for sectors with incomplete data
Non-error sectors with less stored data than expected are now
flagged with a red indicator. It shows a negative figure
indicating the number of bytes short the data field is.

Disks dumped using SAMdisk3 --gaps on systems with unreliable
high-resolution timers may have a large enough positional
errors to cause a few bytes to be clipped from the end of good
sectors. Previously it was only possible to tell this by
viewing the sector data, which showed the stored data size.
The new indicator makes it easy to spot affected images.