Fixed some functionality not working under Windows XP
The stat() function is broken under VS2015 when using the v140_xp
platform toolset. This caused .raw files to be taken as HDD images,
with confusing consequences. Thanks to Gorski for reporting this.

Added FileSize() helper and Win32-specific implementations of
other functions that use stat. Also removed some conditional code
that called the appropriate version of stat, which was messy.
When MS finally fix the bug the new helpers can be simplified.