Fixed track bit length being updated by blank tracks
Scanning a track format that yields no sectors should not update
the bit length, as it was doing for some Amiga tracks.

We'll need to revisit this for multi-format tracks, as scan uses
the most common data encoding but the track doesn't know which
units to store. It's only a guide anyway...

Also added debug logging for detected Amiga sectors.