Added basic USB floppy reading under Linux
- renamed DeviceHDD to BlockDevice
- allowed block devices through to some commands
- added uncached device reading, to avoid read-ahead
- allow override of sector count as valid subset
- added ScsiCmd for shared SG_IO ioctl requests
- changed legacy BLKGETSIZE to new BLKGETSIZE64
- added more PC regular formats
- added SizeToFormat to convert total size to logical format
- CalculateGeometry now supports known floppy formats
- added Sector::set_baddatacrc for manual error flagging
- added page-aligned allocations for non-Win32 platforms
- scan command no longer outputs after being interrupted

USB drives have a limited view of the disk, and make their own
decision about the format of the media. Once made that is used
for the LBA to CHS mapping in any accesses, so only that format
or subsets of it can be accessed.

I may come to regret adding support for this as it'll give some
users false hope that they can magically read any format!