Added preliminary support for modified track detection
If sync is lost between any header and data field on a track we
consider the track to have been modified since original formatting.
Professionally mastered disks are written in a single pass and
shouldn't have write splices in the sector content.

Modified tracks are shown with a '*' symbol after the track number
in scan output, and should be considered a rough guide.

Currently this feature is disabled, due to the way we match headers
from the first revolution with data fields from later revolutions.
It will be re-enabled when headers are corrected to matched with
their correct neighbours.