Fixed DS2 false-positive
400K .dsk images were assumed to be split DSK/DS2 SAM Coupe images,
and would fail when the companion .ds2 image was not found.

The attempt to open the .ds2 image was also giving a confusing file
not found error due to a thrown exception, before a more helpful
error message could be displayed.

We now require both .dsk and .ds2 images to exist and be the correct
size before accepting them as DS2 files.