Add detection and fixing of OperaSoft 32K sectors
OperaSoft titles used a 32K (size=8) sector on cyl 40 for their
copy-protection checks. Data is read from offset 0x512, which
includes the track wrap splice, and written to sector 7. The
protection checks that this data matches or the game crashes.

Unfortunately, CPDRead doesn't store any data for sector size
codes > 6, so this data was usually missing.

SAMdisk now shows a warning if the sector is found to be empty.
Using --fix it will also restore the missing data to fix the
image, as it can be constructed from known filler and the
contents of sector 7 (which is preserved).