Added warning for legacy 6144-byte 8K sectors
Many disks using 8K sectors included 1 or 2 checksum bytes
after the 6K data block, which could be used to validate the
contents in the absence of a FDC checksum. CPDRead only stored
the 6K data area so older images lack the ability to verify
the disk was dumped correctly.

SAMdisk dumps store the full 8K, and attempt to validate the
data using the extra byte(s). This check was silently skipped
if insufficient data was available, but that may cause the user
to think the image was good. We now warn if the data can't
be verified due to the missing checksum.