Added +3 boot patching for Ubi Soft loader
Zombi is a little too quick in using the FDC after switching the
motor back on, which can be a problem for 3.5" floppy drives.

The patch is a generic style that will match any other titles using
the same loader. An attribute clear is replaced by a short delay.

Also expanded the code comments describing the issue.