Fixed CP/M filename display in directory entries
The CPC uses bit 7 in the file extension for hidden and read-only
flags. We were detecting the hidden flag but failing to mask the
extra bit when displaying the names. This was working in v3 but
broken during the transition to v4.

Hidden files are shown in blue, but this isn't visible when the
output isn't an interactive terminal. In this case we now display
a text description of the flags. In either case the description
is sent to the logfile, if enabled.