Limit sector data extent to MFM and FM sectors
The extent calculation relies on specific knowledge of the
ISO MFM/FM formats, and doesn't apply to other encodings.
This was an assumption from the PC-specific SAMdisk v3.

Anything else now returns the natural sector size instead.
This also prevents data_overlap reporting an overlap if the
sectors are quite close together.