Restored support for reading DMK disk images
Images are read into a bitstream to better preserve the source
content. This is fed into the normal bitstream decoder for a
high level view.

The lack of a DAM index in the image does mean their positions
have to be guessed, which can be a bit unreliable for the short
FM address marks. The start of track is also assumed to be the
same format as the first sector.

Added support for RX02 sectors, which are FM up to the start
of the data, then switch to MFM for data and CRC (thanks Chuck!).