Tweaked gap3 size detection rules
Changed to record gap3 size only if it was unshifted within the
bitstream. This should give more consistent values on modified
disks where write splices will cause some variation.

Also removed the requirement that all gap3 sizes need to be the
same (or undefined). We now use the most common gap3 size found,
but favour the previous gap3 if it's still used. If no new gap3
size is detected we assume it's the same as the last track.