Fixed extra zero bit at start of FluxTrackBuffer
Due to precompensation, we need to know the bits either side of
the one we're writing. When the first bit was added we flushed
out the initial zero bit in m_curbit, which wasn't part of the
supplied data. This caused an extra bit to be read back when
decoding the same buffer (visible with scan -v).

The bit doubling of FM data is now performed by TrackBuffer.cpp,
rather than each derived class.

Also renamed some variables to have more logical names.