Ignore bad data fields with different DAM
In some cases we may see a different valid DAM for the same data
field when dealing with damaged disks. Rather than assert on the
unexpected DAM change we can simply ignore it -- the data is likely
to be junk too.

Seen with the disk image image: Reussir Les 4 Operations [A].scp
when scanning using the options: -c39-39 -h0 --pll-adjust=4