Work-around for truncated Amiga writing
We don't yet pad tracks to the full size before writing, so
writing stops after the last byte in the constructed track.
The final data byte wasn't always reliable when written to
the SCP device, so we need to ensure there's some padding.

Moving one of the leading 00 padding bytes to the end works.