Added SuperCard Pro device writing
Currently requires a bitstream or flux source, and will write the
last revolution of stored data. Recognised protections and simple
sector formats are supported.

Added missing 8K sector protection to special bitstream generation.

Also added reading/writing status messages, with a change to flush
the cursor back to the start of the line after each. This ensures
further messages are from the start of line.