Fixed partial status messages being left behind
The status line was not being cleared before new text was shown,
which relied on the new text always being longer than the last
status message, or some of it would be left behind.

We now track which portions of the output are status message,
and clear the line before the next non-status text is shown.
This also allows us to suppress status messages in log output,
since they're generally only of interest to an interactive user.

Thanks to Philippe for identifying this issue.