Allow individual track reads to be uncached
Requests can now be made from the underlying DemandDisk-derived
classes without having to flush the entire disk cache. The new
result written back to the cache to replace the original.

Changed the rpm command to use this. Also changed it to use 2
decimal places, and fix the missing trailing zeros.