Added write support for over-sized error sectors
This adds support for many tracks that contain sectors with
data CRC errors and size codes larger than the space they
occupy. We now attempt to determine the underlying layout
needed, so they can be written to a real track.

If the layout calculation determines a suitable gap3 value
we'll use that when one hasn't been provided on the command-
line or the sector data. Failing that we use a relatively
short gap3 of 25 bytes to give it a reasonable chance to fit.

Also added support for ID header CRC errors, for completeness.