Added tracking of normalised flux data
Raw flux data read directly from a device is not suitable for
writing straight back to the device, as the written positions
won't match the original reversals.

We now track the source of flux data to know whether it's
clean enough to write out. If not we'll convert the bitstream
representation to flux, which also includes an adjustment for
pre-compensation on later tracks.

The clean flux from IPF images and normalised .scp images may
still need a pre-compensation adjustment if it's found to not
be reliable enough.