Added concept of a preferred data source
Rather than always using the native type for an output, the user
can request a particular type is used. For example, writing a flux
image to the SuperCard Pro with --prefer=track causes the sector-
based representation to be converted to bitstream and then flux,
rather than writing the flux directly.

For a flux output, a bitstream source is preferred for unnormalised
flux sources, but normalised flux will be written directly. Using
--prefer=flux will force the available flux to be written regardless
of quality.

Removed --no-flux as it's no longer needed.