Added overhanging sector writing for MFM/FM
When writing to flux floppy devices, we normally write from
index to index. If the last data field overhangs into the index
and the source data is bitstream or flux we'd stop writing in
the middle of the data field, generating a data error on read.

We now detect good MFM/FM sectors that overhang the index and
mark the bitstream splice point where the data field ends. This
allows us to add enough data to stop writing at that position
instead of the index, preserving the sector.

This will need extending to other bitstream decoders (Amiga in
particular), and the data extension moved out of the SCP writer
into a function that can be shared with other writers.