Improved raw file handling and overrides
Raw files are matched against the list of known file sizes, to
give a starting geometry. User overrides are then applied, and if
the geometry matches the file size and is valid, it's accepted.

If matched and none of the cyls/heads/sectors/size are changed,
the user is warned that the default may not be correct, to avoid
accidental use of unwanted formats.

If the file size isn't recognised, the user can still specify
geometry values that match the file size to have it accepted.
The default is 80/2/0/512, and only the changed parts need to
be specified.

Sector size/count are now linked, as are cyls/heads. If only one
of the values is changed, the other is calculated such that the
original file size is preserved.