Added hard-sector count option
The new hard-sector count option (-H or --hard-sectors) specifies
how many neighbouring index holes to combine to give a full track.

This is currently only supported for KryoFlux images, but can be
expanded to other formats as needed. The SuperCard Pro firmware
appears to limit the revolution count to 5, which may be an issue.