Added support for PDI disk images
These are RLE-compressed DOS disk images (720K and 1.44M).

The original file format document (PDITYPE1.TXT) couldn't be found,
so format details were taken from the existing PDI2IMA utility.

Also fixed a bug in one of the MemFile::read overloads, which was
failing to advance the iterator for the returned data.