Second attempt at fixing Victor false-positives
The previous change failed because a rogue header is enough to make
us think we found something. That only requires a SYNC followed by
an IDAM of 07, and a checksum of zero. Once that matches we'll
report unrecognised address marks after every potential SYNC.

The unrecognised AM messages are now limited to debug mode only,
and disabled by default. Apple and GCR logging has been reduced
too, since it produces hundreds of lines of output. We might need
a debug verbosity level to handle both cases.

We now only accept the track as Victor format if we found a header
and data field close enough to give a complete sector.