Fixed Prehistoric format check
The Prehistoric tracks generated for the built-in test disk didn't pass
the IsPrehistorikTrack test. The check has been relaxed to remove the
cylinder check, and the sector size is determined by data error status.

Added a basic template for the alternate Prehistoric was seen on some
Titus The Fox [CPC] disks, which have unused KBI-19 sectors at the end
of the track. GeneratePrehistorikTrack still needs enhancing for this.