Restored --tty terminal detection override
This forces ANSI escape sequences to be output, even if the terminal isn't
detected as interactive. This is useful for piping long output to less -R

The option doesn't work on Windows, which still uses the console API. Though
Windows 10 can now support ANSI escape sequences so it may in the future.